Update Hooks.cs
Me and my friend were trying were trying this mod out! Pretty hectic! Some suggestion here. Stretch the the difficulty out a bit more so it doesn't catch up to players in a bad way and bump up the base move speed by a lot! If you're not for the change to the base move speed, of the difficulty timer, maybe you could make it so those parameters and others are available for changing in r2modman's config edit option. At least just the base move speed! Please and thank you!